+++ ioflo/__metadata__.py
@@ -2,7 +2,7 @@
 Ioflo package metadata
 '''
 
-__version_info__ = (1, 4, 9)
+__version_info__ = (1, 5, 0)
 __version__ = '{0}.{1}.{2}'.format(*__version_info__)
 __author__ = "Samuel M. Smith"
 __license__ = "Apache 2.0"